Rydder opp i navn og plassering av funksjon
