+++ src/felleskomponenter/oppsummering/oppsummering.tsx
@@ -80,7 +80,10 @@ const Oppsummering = ({
   const disableEndreKnapp = behandlingsStatusMedBegrensetRettigheter.includes(oppsummering.behandlingsstatus.kode);
   const erLitenSkjerm = Utils.mediaQuery.useMediaQuery({ maxWidth: 1440 });
 
-  const erSed = behandlingstype?.kode === MKV.Koder.behandlinger.behandlingstyper.SED;
+  const erSed =
+    behandleAlleSakerToggle === "enabled"
+      ? MKVUtils.erBehandlingAvSed(behandlingstema?.kode, fagsak.sakstype?.kode)
+      : behandlingstype?.kode === MKV.Koder.behandlinger.behandlingstyper.SED;
   const erTrygdeavtale = sakstype?.kode === MKV.Koder.sakstyper.TRYGDEAVTALE;
   const hovedpartErVirksomhet = hovedpartRolle === MKV.Koder.aktoersroller.VIRKSOMHET;
