+++ src/melosyskodeverk/kodekombinasjoner.js
@@ -22,29 +22,6 @@ const bestemmelserIkkeRelevanteForUnntak = [
   MKV.Koder.lovvalgsbestemmelser.lovvalgbestemmelser_987_2009.FO_987_2009_ART14_11,
 ];
 
-export const standardBehandlingsTemaMedBegrensetRettigheter = [
-  MKV.Koder.behandlinger.behandlingstema.REGISTRERING_UNNTAK_NORSK_TRYGD_UTSTASJONERING,
-  MKV.Koder.behandlinger.behandlingstema.REGISTRERING_UNNTAK_NORSK_TRYGD_ØVRIGE,
-  MKV.Koder.behandlinger.behandlingstema.BESLUTNING_LOVVALG_NORGE,
-  MKV.Koder.behandlinger.behandlingstema.BESLUTNING_LOVVALG_ANNET_LAND,
-];
-
-export const utvidetBehandlingsTemaMedBegrensetRettigheter = [
-  ...standardBehandlingsTemaMedBegrensetRettigheter,
-  MKV.Koder.behandlinger.behandlingstema.ANMODNING_OM_UNNTAK_HOVEDREGEL,
-];
-
-export const erEOS = (sakstype) => {
-  return sakstype === MKV.Koder.sakstyper.EU_EOS;
-};
-
-// Denne kan nok flyttes inn i endreBehandlingModal siden den kun brukes der etter melosys.behandle_alle_saker fjernes
-export const erBehandlingstemaMedBegrensetRettigheter = (behandlingstema, sakstype) => {
-  return erEOS(sakstype)
-    ? utvidetBehandlingsTemaMedBegrensetRettigheter.includes(behandlingstema)
-    : standardBehandlingsTemaMedBegrensetRettigheter.includes(behandlingstema);
-};
-
 const kodeverkComparator = ({ kode: k1 }, { kode: k2 }) => k1.localeCompare(k2);
 
 export const unntaksbestemmelser = Utils._uniqBy(
